+++ module-check/src/main/clojure/clojure/core/typed/check_form_clj.clj
@@ -17,10 +17,14 @@
 
 (defn check-form-info
   [form & opt]
-  (apply chk-form/check-form-info (config-map)
-         form opt))
+  (let [config (config-map)]
+    (impl/with-full-impl (:impl config)
+      (apply chk-form/check-form-info config
+             form opt))))
 
 (defn check-form*
   [form expected type-provided?]
-  (chk-form/check-form* (config-map)
-    form expected type-provided?))
+  (let [config (config-map)]
+    (impl/with-full-impl (:impl config)
+      (chk-form/check-form* config
+        form expected type-provided?))))