+++ module-rt/src/main/clojure/clojure/core/typed/util_vars.clj
@@ -34,7 +34,6 @@
   nil)
 
 (def ^:skip-wiki ^:dynamic *currently-loading* false)
-(defonce ^{:doc "True during type checking."} ^:dynamic *checking* nil)
 (defonce ^{:doc "Internal use only"} ^:skip-wiki ^:dynamic *already-collected* nil)
 (defonce ^{:doc "Internal use only"} ^:skip-wiki ^:dynamic *already-checked* nil)
 (defonce ^{:doc "Internal use only"} ^:skip-wiki ^:dynamic *delayed-errors* nil)