+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Thanks to Allen Rohner who provided patches for this release.
 
 - Fix CTYP-215
 - Fix CTYP-174
+- Fix CTYP-181
+  - better types for primitive casts
 
 # 0.2.88 - 18 May 2015