+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Thanks to Allen Rohner who provided patches for this release.
 - Fix CTYP-181
   - better types for primitive casts
 - Fix CTYP-170
+- Fix CTYP-200
 
 # 0.2.88 - 18 May 2015